# be/docs/swagger.yaml
basePath: /
definitions:
...
  api.OllamaRequest:
    properties:
      prompt:
        type: string
    type: object
  api.OllamaResponse:
    properties:
      error:
        type: string
      response:
        type: string
    type: object
...
host: localhost:1971
info:
  contact:
    email: support@example.com
    name: API Support
  description: This is the API documentation for the ρBee (rhobee) backend server.
    It provides endpoints for user authentication, navigation, file management, and
    integration with Ollama AI services.
  license:
    name: Apache 2.0
    url: https://opensource.org/licenses/Apache-2.0
  termsOfService: http://swagger.io/terms/
  title: ρBee (rhobee) API
  version: "1.0"
paths:
...
  /ollama:
    post:
      consumes:
      - application/json
      description: Sends a prompt to the Ollama API and returns the generated response
      parameters:
      - description: Request body containing the prompt
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.OllamaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ollama response
          schema:
            $ref: '#/definitions/api.OllamaResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/api.OllamaResponse'
        "503":
          description: Ollama service not configured
          schema:
            $ref: '#/definitions/api.OllamaResponse'
      security:
      - BearerAuth: []
      summary: sends a prompt to Ollama and returns the response
      tags:
      - ollama
  /ollama/default-page:
    get:
      description: Returns the default page response generated by Ollama in the requested
        language
      parameters:
      - default: en
        description: Language tag (e.g., en, it, fr, de)
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Default page response
          schema:
            $ref: '#/definitions/api.OllamaResponse'
        "503":
          description: Ollama service not configured
          schema:
            $ref: '#/definitions/api.OllamaResponse'
      summary: returns default page response generated by Ollama
      tags:
      - ollama
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...